refactor: price 컴포넌트 계층화
