+++ client/src/components/atoms/price/PriceAndQuantity.tsx
@@ -1,5 +1,5 @@
-import { TElementProps } from "../../types";
-import { convertToNumberFormat } from "../../utils";
+import { TElementProps } from "../../../types";
+import { convertToNumberFormat } from "../../../utils";
 
 interface IProps extends TElementProps<HTMLSpanElement> {
   price: number;
