+++ client/src/components/atoms/price/index.ts
@@ -1,0 +1,2 @@
+export { default as Price } from "./Price";
+export { default as PriceAndQuantity } from "./PriceAndQuantity";
